refactor(cascade): Stage C.2 — split core/cascade.py 24KB → package, all files ≤12KB
Splits the 24 KB monolithic core/cascade.py into a 3-module package
along the Phase C / Phase D / shared-helpers seam. Resolves
PLATFORM_READINESS Gate v0.3 dimension A for this file (2nd of five
violations) — first was wiki_generator.py 51 KB in PR #427.

## What changed

- core/cascade.py (24 KB, 588 lines, 11 module-level functions, no
  class) → core/cascade/ package preserving the same external surface.
- No function signatures, return shapes, or behaviour changed. Split
  is purely textual.

## File sizes (post-split)

  __init__.py     2.0 KB   facade, re-exports
  _helpers.py     2.4 KB   uuid regex + frontmatter read/write +
                           entity-file iterator (shared by C + D)
  _delete.py     11.3 KB   Phase C — 5 functions
  _modify.py     11.9 KB   Phase D — 3 functions

All four files comfortably under the 20 KB gate.

## Function placement

Phase C (delete) → _delete.py:
- cascade_remove_doc_from_sources (sources sweep)
- find_orphan_entities (incoming=0 + remaining-relations=0 rule)
- find_doc_entity_path
- backup_upload_file (.deleted/ archive)
- cascade_delete_upload (top-level pipeline)

Phase D (modify) → _modify.py:
- load_extraction_sidecar
- diff_triples (added/removed/kept classification)
- cascade_modify_doc (top-level pipeline; reuses
  cascade_remove_doc_from_sources + find_orphan_entities +
  backup_upload_file from _delete.py)

Shared utilities → _helpers.py:
- _UUID_PREFIX_RE / _FM_SPLIT_RE
- strip_uuid_prefix
- _read_frontmatter / _write_frontmatter (cascade-local; distinct
  from the WikiGenerator method of the same name)
- _iter_entity_files

## External surface preserved

All current import paths work byte-identical:

- from core.cascade import cascade_delete_upload — server_llmwiki.py:5215
- from core.cascade import cascade_modify_doc — server_llmwiki.py:5306
- from core.cascade import (...) — tests/test_phase_c_cascade.py:35,
  tests/test_phase_d_modify_cascade.py:35,
  tests/test_phase_e_graph_editor.py:394

The package __init__.py re-exports every name that was at module
level in the old file, including the underscore-prefixed helpers that
tests may patch.

## Verification

- tests/test_phase_c_cascade.py + test_phase_d_modify_cascade.py +
  test_phase_e_graph_editor.py — 75/75 pass
- Full pytest suite matching CI's ignore list: 1590 passed + 824
  subtests passed. Zero regression.
- scripts/dogfood_check.py — 4/4 plugin invariants still PASS.
- ruff check core/cascade/ — clean.

## Stage C progress

- C.1 (wiki_generator.py 51 KB) — ✅ PR #427
- C.2 (cascade.py 24 KB) — ✅ this PR
- C.3 (character_profile.py 24 KB) — pending
- C.4 (security_layer.py 23 KB) — pending (after Stage D.1 JSONL writers)
- C.5 (graph_editor.py 20 KB boundary) — pending; reassess after C.4

## Out of scope

- Phase E (graph editor) cascade — already lives in
  core/graph_editor.py, unchanged here.
- Behaviour-level changes to cascade semantics — this PR is pure
  refactor.
